+++ examples/actix3_example/core/src/query.rs
@@ -11,9 +11,9 @@ impl Query {
     /// If ok, returns (post models, num pages).
     pub async fn find_posts_in_page(
         db: &DbConn,
-        page: usize,
-        posts_per_page: usize,
-    ) -> Result<(Vec<post::Model>, usize), DbErr> {
+        page: u64,
+        posts_per_page: u64,
+    ) -> Result<(Vec<post::Model>, u64), DbErr> {
         // Setup paginator
         let paginator = Post::find()
             .order_by_asc(post::Column::Id)
